Add absolute and relative path reference operators
These operators let you insert the absolute or relative path of a fragment's
file into the woven output. I imagine this will be most useful for applications
where the fragment sources and woven sources are stored together and referenced
as a unit. For example, this could be used to build the basis of a static site
generator, automatically inserting links between pages into the woven output.